S06L074-S06L074: Authentification Centralized

+++ prisma/src/resolvers/mutation.js
@@ -1,5 +1,6 @@
 import bcrypt from 'bcryptjs'
 import jwt from 'jsonwebtoken'
+import getUserId from '../utils/getUserId'
 
 const Mutation = {
     async createUser(parent, {data}, { prisma }, info) {
@@ -21,24 +22,53 @@ const Mutation = {
         }
     },
 
-    async deleteUser(parent, args, { prisma }, info) {
+    async login(parent, {data}, {prisma}, info) {
+        const user = await prisma.query.user({
+            where: {
+                email: data.email
+            }
+        })
+
+        if(!user) {
+            throw new Error('Unable to login')
+        }
+
+        const isMatch = await bcrypt.compare(data.password, user.password)
+
+        if(!isMatch) {
+            throw new Error('Unable to login')
+        }
+
+        return {
+            user,
+            token: jwt.sign({userId: user.id}, 'token')
+        }
+    },
+
+    async deleteUser(parent, args, {prisma, request}, info) {
+        const userId = getUserId(request)
+
         return await prisma.mutation.deleteUser({
             where: {
-                id:args.id
+                id:userId
             }
         }, info)
     },
 
-    async updateUser(parent, args, { prisma }, info) {  // args or {id, data} is the same
+    async updateUser(parent, args, {prisma, request}, info) {  // args or {id, data} is the same
+        const userId = getUserId(request)
+
         return await prisma.mutation.updateUser({
             where: {
-                id: args.id
+                id: userId
             },
             data: args.data
         }, info)
     },
 
-    async createPost(parent, args, { prisma }, info) {
+    async createPost(parent, args, {prisma, request}, info) {
+        const userId = getUserId(request)
+
         return await prisma.mutation.createPost({ 
             data:{
                 title: args.data.title,
@@ -46,14 +76,26 @@ const Mutation = {
                 published: args.data.published,
                 author: {
                     connect: {
-                        id: args.data.author
+                        id: userId
                     }
                 }
             } 
         }, info)
     },
 
-    async deletePost(parent, args, { prisma }, info) {
+    async deletePost(parent, args, {prisma, request}, info) {
+        const userId = getUserId(request)
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+        
+        if(!postExists) {
+            throw new Error('Unable to find the post')
+        }
+
         return await prisma.mutation.deletePost({
             where: {
                 id:args.id
